invoker/requirements.txt
functions-framework==3.*
google-auth==2.*
requests>=2.31.0
